% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic_plot.R
\name{cic_plot}
\alias{cic_plot}
\title{Estimate a changes-in-changes model with multiple periods and cohorts}
\usage{
cic_plot(
  object,
  xlab = NULL,
  ylab = "QTE",
  ylim = NULL,
  size = 2,
  es_type = c(NULL, "aggregated", "for_quantiles", "for_periods"),
  perc_plot = seq(0.1, 0.9, 0.1),
  legend_title = "Percentiles"
)
}
\arguments{
\item{object}{An \code{ecic} object.}

\item{xlab}{Alternative x-axis label.}

\item{ylab}{Alternative y-axis label.}

\item{ylim}{Define the y-axis limits.}

\item{size}{Size of the point estimates.}

\item{es_type}{If an event study was estimated with \code{ecic}, you can choose the
styple of the ES plot. "aggregated" puts everything in one plot. "for_quantiles"
generates one plot for each percentile. "for_periods" generates one plot for each period.}

\item{perc_plot}{Which percentiles to plot.}

\item{legend_title}{Change the title of the legend.}
}
\value{
An \code{ggplot2} object.
}
\description{
Estimate a changes-in-changes model with multiple periods and cohorts
}
