% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic_summary.R
\name{cic_summary}
\alias{cic_summary}
\title{Estimate a changes-in-changes model with multiple periods and cohorts}
\usage{
cic_summary(object)
}
\arguments{
\item{object}{An \code{ecic} object.}
}
\value{
An \code{ecic_res} object.
}
\description{
Estimate a changes-in-changes model with multiple periods and cohorts
}
