% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Simulated sample data}
\format{
\subsection{\code{dat}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{countyreal}{Unit ID}
\item{first.treat}{Cohort}
\item{year}{Period}
\item{time_to_treat}{Period - Cohort}
\item{lemp}{dependent variable}
...
}
}
}
\source{
Simulation data
}
\usage{
dat
}
\description{
A simulated sample panel data with heterogenous treatment effects between
cohorts and groups.
}
\keyword{datasets}
