% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ecic.R
\name{print.ecic}
\alias{print.ecic}
\title{Print ecic objects}
\usage{
\method{print}{ecic}(x, ..., details = FALSE)
}
\arguments{
\item{x}{An \code{ecic} object.}

\item{...}{further arguments}

\item{details}{logical. Set to TRUE to print background information for
every bootstrap run and Changes-in-Changes model.}
}
\value{
An \code{ecic} print object.
}
\description{
Prints an \code{ecic} model while making attributes invisible.
}
