% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ecic.R
\name{summary.ecic}
\alias{summary.ecic}
\title{Summary for a changes-in-changes regression with multiple periods and cohorts}
\usage{
\method{summary}{ecic}(object, ...)
}
\arguments{
\item{object}{An \code{ecic} object.}

\item{...}{further arguments.}
}
\value{
An \code{ecic_table} object.
}
\description{
Summarizes an \code{ecic} object by aggregating the bootstrap runs.
Works also in an event-study fashion.
}
