% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ecic.R
\name{summary_ecic}
\alias{summary_ecic}
\title{Summary for a changes-in-changes model with multiple periods and cohorts}
\usage{
summary_ecic(object)
}
\arguments{
\item{object}{An \code{ecic} object.}
}
\value{
An \code{ecic_res} object.
}
\description{
Summarizes an \code{ecic} model by aggregating the bootstrap runs.
Works also in an event-study fashion.
}
