% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecic.R
\name{ecic}
\alias{ecic}
\title{Estimate a changes-in-changes model with multiple periods and cohorts}
\usage{
ecic(
  yvar = NULL,
  gvar = NULL,
  tvar = NULL,
  ivar = NULL,
  dat = NULL,
  myProbs = seq(0.1, 0.9, 0.1),
  nMin = 40,
  boot = c("no", "normal", "weighted"),
  nReps = 1,
  weight_n0 = c("n1", "n0"),
  weight_n1 = c("n1", "n0"),
  quant_algo = 1,
  no_imp = 1e+05,
  es = F,
  periods_es = 6,
  short_output = T,
  save_to_disk = F,
  myFolder = "temp",
  nCores = 1
)
}
\arguments{
\item{yvar}{Dependent variable.}

\item{gvar}{Group variable. Can be either a string (e.g., "first_treated")
or an expression (e.g., first_treated). In a staggered treatment setting,
the group variable typically denotes treatment cohort.}

\item{tvar}{Time variable. Can be a string (e.g., "year") or an expression
(e.g., year).}

\item{ivar}{Index variable. Can be a string (e.g., "country") or an
expression (e.g., country).}

\item{dat}{The data set.}

\item{myProbs}{Quantiles that the quantile treatment effects should be calculated for.}

\item{nMin}{Minimum observations per groups. Small groups are deleted.}

\item{boot}{Bootstrap. Resampling is done over the entire data set ("normal"),
but might be weighted by period-cohort size ("weighted").}

\item{nReps}{Number of bootstrap replications.}

\item{weight_n0}{Weight for the aggregation of the CDFs in the control group.
\code{n1} uses cohort sizes (Alternative: \code{n0}).}

\item{weight_n1}{Weight for the aggregation of the CDFs in the treatment group.
\code{n1} uses cohort sizes (Alternative: \code{n0}).}

\item{quant_algo}{Quantile algorithm (see Wikipedia for definitions).}

\item{no_imp}{Grid size for the imputation of the empirical CDF.}

\item{es}{Event Study (Logical). If TRUE, a QTE is estimated for each period.}

\item{periods_es}{Periods of the event study.}

\item{short_output}{Only reports essential results.}

\item{save_to_disk}{Logical. If TRUE, results are temporarily saved, reduces the
RAM needed.}

\item{myFolder}{Location of the temporary files.}

\item{nCores}{Number of cores used.}
}
\value{
An \code{ecic} object.
}
\description{
Calculates a changes-in-changes model as in Athey and Imbens (2006) for multiple periods and cohorts.
}
\examples{
\donttest{
# Load some sample data
data(dat, package = "ecic")

# Estimate a basic model
mod_res =
  summary_ecic(
  ecic(
    yvar  = lemp,         # dependent variable
    gvar  = first.treat,  # group indicator
    tvar  = year,         # time indicator
    ivar  = countyreal,   # unit ID
    dat   = dat,          # dataset
    boot  = "weighted",   # bootstrap proceduce ("no", "normal", or "weighted")
    nReps = 20            # number of bootstrap runs
  )
  )
  
# Basic Plot
plot_ecic(mod_res)
}
}
\references{
Athey, Susan and Guido W. Imbens (2006). \cite{Identification and Inference in
Nonlinear Difference-in-Differences Models}.
\doi{10.1111/j.1468-0262.2006.00668.x}
}
